Extract the problem/result display properties to the problemcalculationsviewmodel
