+++ .gitignore
@@ -10,3 +10,4 @@
 3DPoC/POC3D/POC3D.ViewModel/obj/
 3DPoC/POC3D/POC3D.ViewModel.Tests/bin/
 3DPoC/POC3D/POC3D.ViewModel.Tests/obj/
+3DPoC/POC3D/POC3D.ViewModel/bin/
