+++ 3DPoC/POC3D/POC3D.ViewModel/Calculations/ProblemCalculationViewModel.cs
@@ -1,4 +1,6 @@
 ﻿using POC3D.ViewModel.Base;
+using System.Threading.Tasks;
+using System.Windows;
 
 namespace POC3D.ViewModel.Calculations
 {
@@ -14,6 +16,11 @@ namespace POC3D.ViewModel.Calculations
         private NumericMatrix _fullSolvedDisplacementsVector;
         private NumericMatrix _solvedReactionForces;
         private bool? _canBeSolved;
+
+
+        private bool _displacementAnimation;
+        private double _displacementsMultiplier = 1;
+        private bool _showProblem = true;
 
         public ProblemCalculationViewModel(ProblemViewModel problemViewModel)
         {
@@ -64,6 +71,89 @@ namespace POC3D.ViewModel.Calculations
             OnPropertyChanged(nameof(FullSolvedDisplacementsVector));
             OnPropertyChanged(nameof(SolvedReactionForces));
             OnPropertyChanged(nameof(CanBeSolved));
+        }
+
+        public double DisplacementsMultiplier
+        {
+            get => _displacementsMultiplier;
+            set
+            {
+                _displacementsMultiplier = value;
+                if (!_showProblem)
+                    UpdateDisplacementsInNodes();
+                OnPropertyChanged(nameof(DisplacementsMultiplier));
+            }
+        }
+
+        public bool DisplacementAnimation
+        {
+            get => _displacementAnimation;
+            set
+            {
+                _displacementAnimation = value;
+
+                if (_displacementAnimation) Application.Current.Dispatcher.InvokeAsync(RunAnimation);
+            }
+        }
+
+        public bool ShowProblem
+        {
+            get => _showProblem;
+            set
+            {
+                _showProblem = value;
+                if (!_showProblem)
+                    UpdateDisplacementsInNodes();
+                OnPropertyChanged(nameof(ShowProblem));
+            }
+        }
+
+        private void UpdateDisplacementsInNodes()
+        {
+            var index = 0;
+
+            foreach (var node in _problemViewModel.Nodes)
+            {
+                if (node.IsFixed)
+                {
+                    node.DisplacementX = 0;
+                    node.DisplacementY = 0;
+                    node.DisplacementZ = 0;
+                }
+                else
+                {
+                    node.DisplacementX = SolvedDisplacementsVector[index + 0, 0] * DisplacementsMultiplier;
+                    node.DisplacementY = SolvedDisplacementsVector[index + 1, 0] * DisplacementsMultiplier;
+                    node.DisplacementZ = SolvedDisplacementsVector[index + 2, 0] * DisplacementsMultiplier;
+                    index = index + 3;
+                }
+            }
+        }
+
+
+        private async Task RunAnimation()
+        {
+            const int delay = 100;
+            var savedMultiplier = DisplacementsMultiplier;
+
+            var delta = savedMultiplier / 10.0;
+
+            while (DisplacementAnimation)
+            {
+                for (var i = 0; i < 10; i++)
+                {
+                    DisplacementsMultiplier -= delta;
+                    await Task.Delay(delay);
+                }
+
+                for (var i = 0; i < 10; i++)
+                {
+                    DisplacementsMultiplier += delta;
+                    await Task.Delay(delay);
+                }
+            }
+
+            DisplacementsMultiplier = savedMultiplier;
         }
     }
 }
