+++ 3DPoC/POC3D/POC3D.ViewModel/ProblemViewModel.cs
@@ -15,14 +15,9 @@ namespace POC3D.ViewModel
     public class ProblemViewModel : Observable
     {
         private readonly IModelProblem _modelProblem;
-
-        private bool _displacementAnimation;
-        private double _displacementsMultiplier = 1;
         private ElementViewModel _selectedElement;
         private ForceViewModel _selectedForce;
         private NodeViewModel _selectedNode;
-
-        private bool _showProblem = true;
 
         public ProblemViewModel()
         {
@@ -131,67 +126,7 @@ namespace POC3D.ViewModel
 
         public int NumberOfDirichletBoundaryConditions => Nodes.Count(x => x.IsFixed);
 
-        public bool ShowProblem
-        {
-            get => _showProblem;
-            set
-            {
-                _showProblem = value;
-                if (!_showProblem)
-                    UpdateDisplacementsInNodes();
-                OnPropertyChanged(nameof(ShowProblem));
-            }
-        }
-
-        public double DisplacementsMultiplier
-        {
-            get => _displacementsMultiplier;
-            set
-            {
-                _displacementsMultiplier = value;
-                if (!_showProblem)
-                    UpdateDisplacementsInNodes();
-                OnPropertyChanged(nameof(DisplacementsMultiplier));
-            }
-        }
-
-        public bool DisplacementAnimation
-        {
-            get => _displacementAnimation;
-            set
-            {
-                _displacementAnimation = value;
-
-                if (_displacementAnimation) Application.Current.Dispatcher.InvokeAsync(RunAnimation);
-            }
-        }
-
         public ProblemCalculationViewModel ProblemCalculationViewModel { get; }
-
-        private async Task RunAnimation()
-        {
-            const int delay = 100;
-            var savedMultiplier = DisplacementsMultiplier;
-
-            var delta = savedMultiplier / 10.0;
-
-            while (DisplacementAnimation)
-            {
-                for (var i = 0; i < 10; i++)
-                {
-                    DisplacementsMultiplier -= delta;
-                    await Task.Delay(delay);
-                }
-
-                for (var i = 0; i < 10; i++)
-                {
-                    DisplacementsMultiplier += delta;
-                    await Task.Delay(delay);
-                }
-            }
-
-            DisplacementsMultiplier = savedMultiplier;
-        }
 
         public NodeViewModel AddNode(Point3D point)
         {
@@ -302,27 +237,6 @@ namespace POC3D.ViewModel
             OnPropertyChanged(nameof(NumberOfNodes));
             OnPropertyChanged(nameof(NumberOfElements));
             OnPropertyChanged(nameof(NumberOfDirichletBoundaryConditions));
-        }
-
-        private void UpdateDisplacementsInNodes()
-        {
-            var index = 0;
-            foreach (var node in Nodes)
-            {
-                if (node.IsFixed)
-                {
-                    node.DisplacementX = 0;
-                    node.DisplacementY = 0;
-                    node.DisplacementZ = 0;
-                }
-                else
-                {
-                    node.DisplacementX = ProblemCalculationViewModel.SolvedDisplacementsVector[index + 0, 0] * DisplacementsMultiplier;
-                    node.DisplacementY = ProblemCalculationViewModel.SolvedDisplacementsVector[index + 1, 0] * DisplacementsMultiplier;
-                    node.DisplacementZ = ProblemCalculationViewModel.SolvedDisplacementsVector[index + 2, 0] * DisplacementsMultiplier;
-                    index = index + 3;
-                }
-            }
         }
 
         private MaterialViewModel GetOrInitMaterialViewModel(IModelMaterial modelMaterial)
